feat: Implement full CRUDs APIs for TrainingProgram

+++ backend/TrainingProgram.Api/Controllers/MacrocyclesController.cs
@@ -1,16 +1,20 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using TrainingProgram.Api.Data;
+using TrainingProgram.Api.Dtos;
 using TrainingProgram.Api.Models;
 
 namespace TrainingProgram.Api.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Authorize]
     public class MacrocyclesController : ControllerBase
     {
         private readonly AppDbContext _context;
@@ -20,10 +24,19 @@ namespace TrainingProgram.Api.Controllers
             _context = context;
         }
 
+        // This read the Token and return his ID
+        private string GetCurrentUserId()
+        {
+            return User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        }
+
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Macrocycle>>> GetMacrocycles()
         {
+            var userId = GetCurrentUserId();
+
             var macrocycles = await _context.Macrocycles
+                                            .Where(m => m.UserId == userId)
                                             .Include(m => m.Microcycles)
                                             .ToListAsync();
 
@@ -33,32 +46,55 @@ namespace TrainingProgram.Api.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Macrocycle>> GetMacrocycle(int id)
         {
+            var userId = GetCurrentUserId();
+
             var macrocycle = await _context.Macrocycles
                                             .Include(m => m.Microcycles)
                                             .ThenInclude(micro => micro.TrainingSessions)
-                                            .ThenInclude(train => train.Activities)
+                                            .ThenInclude(session => session.Activities)
                                             .FirstOrDefaultAsync(m => m.Id == id);
             if (macrocycle == null) return NotFound();
+
+            if (macrocycle.UserId != userId) return Forbid(); // 403 Forbidden
+
 
             return Ok(macrocycle);
         }
 
         [HttpPost]
-        public async Task<ActionResult<Macrocycle>> CreateMacrocycle(Macrocycle macrocycle)
+        public async Task<ActionResult<Macrocycle>> CreateMacrocycle(CreateMacrocycleDto macrocycle)
         {
-            _context.Macrocycles.Add(macrocycle);
+            var userId = GetCurrentUserId();
+
+            var newMacrocycle = new Macrocycle
+            {
+                Year = macrocycle.Year,
+                TeamName = macrocycle.TeamName,
+                CoachName = macrocycle.CoachName,
+                UserId = userId
+            };
+
+            _context.Macrocycles.Add(newMacrocycle);
             await _context.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(GetMacrocycle), new { id = macrocycle.Id }, macrocycle);
+            return CreatedAtAction(nameof(GetMacrocycle), new { id = newMacrocycle.Id }, newMacrocycle);
 
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult> UpdateMacrocycle(Macrocycle macrocycle, int id)
+        public async Task<ActionResult> UpdateMacrocycle(UpdateMacrocycleDto macrocycle, int id)
         {
-            if (id != macrocycle.Id) return BadRequest();
+            var macrocycleDb = await _context.Macrocycles.FindAsync(id);
 
-            _context.Entry(macrocycle).State = EntityState.Modified;
+            if (macrocycleDb == null) return NotFound();
+
+            var userId = GetCurrentUserId();
+
+            if (macrocycleDb.UserId != userId) return Forbid(); // 403 Forbidden
+
+            macrocycleDb.Year = macrocycle.Year;
+            macrocycleDb.TeamName = macrocycle.TeamName;
+            macrocycleDb.CoachName = macrocycle.CoachName;
 
             try
             {
@@ -86,7 +122,11 @@ namespace TrainingProgram.Api.Controllers
         {
             var macrocycle = await _context.Macrocycles.FindAsync(id);
 
+
             if (macrocycle == null) return NotFound();
+
+            var userId = GetCurrentUserId();
+            if (macrocycle.UserId != userId) return Forbid(); // 403 Forbidden
 
             _context.Macrocycles.Remove(macrocycle);
             await _context.SaveChangesAsync();
